requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.0
tweepy>=4.14.0
aiohttp>=3.8.0
openai>=1.0.0
discord.py>=2.0.0
python-telegram-bot>=20.0
pysui>=0.80.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API interaction
requests>=2.31.0
websockets>=11.0.0

# Utility
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0
loguru>=0.7.0
textblob==0.17.1

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
Sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
sphinx-autodoc-typehints>=1.23.0
sphinx-multiversion>=0.2.4 